refactor: Migrate workout contexts to hooks structure
- Refactored CurrentWorkoutInstanceContext and GeneratedWorkoutContext to utilize a centralized hooks structure, enhancing organization and maintainability.
- Updated imports across various components to use hooks from the new hooks directory, improving clarity and reducing redundancy.
- Removed obsolete context definitions and custom hooks related to workout instances and generated workouts, streamlining the codebase for better performance and readability.
