+++ src/hooks/index.ts
@@ -28,3 +28,5 @@ export { usePusherBeamsInterests } from "./usePusherBeamsInterests";
 export { usePusherEvent } from "./usePusherEvent";
 export * from "./useWorkoutFeedback";
 export * from "./useTrainerPersona";
+export * from "./useGeneratedWorkouts";
+export * from "./useCurrentWorkoutInstance";
